requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
decorator==5.1.1
distro==1.9.0
Django==5.0.7
executing==2.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab_widgets==3.0.11
langchain==0.2.11
langchain-core==0.2.23
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
matplotlib-inline==0.1.7
multidict==6.0.5
numpy==1.26.4
openai==1.37.0
orjson==3.10.6
packaging==24.1
parso==0.8.4
pexpect==4.9.0
playsound==1.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyobjc==10.3.1
pyobjc-core==10.3.1
pyobjc-framework-Accessibility==10.3.1
pyobjc-framework-Accounts==10.3.1
pyobjc-framework-AddressBook==10.3.1
pyobjc-framework-AdServices==10.3.1
pyobjc-framework-AdSupport==10.3.1
pyobjc-framework-AppleScriptKit==10.3.1
pyobjc-framework-AppleScriptObjC==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-AppTrackingTransparency==10.3.1
pyobjc-framework-AudioVideoBridging==10.3.1
pyobjc-framework-AuthenticationServices==10.3.1
pyobjc-framework-AutomaticAssessmentConfiguration==10.3.1
pyobjc-framework-Automator==10.3.1
pyobjc-framework-AVFoundation==10.3.1
pyobjc-framework-AVKit==10.3.1
pyobjc-framework-AVRouting==10.3.1
pyobjc-framework-BackgroundAssets==10.3.1
pyobjc-framework-BrowserEngineKit==10.3.1
pyobjc-framework-BusinessChat==10.3.1
pyobjc-framework-CalendarStore==10.3.1
pyobjc-framework-CallKit==10.3.1
pyobjc-framework-CFNetwork==10.3.1
pyobjc-framework-Cinematic==10.3.1
pyobjc-framework-ClassKit==10.3.1
pyobjc-framework-CloudKit==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Collaboration==10.3.1
pyobjc-framework-ColorSync==10.3.1
pyobjc-framework-Contacts==10.3.1
pyobjc-framework-ContactsUI==10.3.1
pyobjc-framework-CoreAudio==10.3.1
pyobjc-framework-CoreAudioKit==10.3.1
pyobjc-framework-CoreBluetooth==10.3.1
pyobjc-framework-CoreData==10.3.1
pyobjc-framework-CoreHaptics==10.3.1
pyobjc-framework-CoreLocation==10.3.1
pyobjc-framework-CoreMedia==10.3.1
pyobjc-framework-CoreMediaIO==10.3.1
pyobjc-framework-CoreMIDI==10.3.1
pyobjc-framework-CoreML==10.3.1
pyobjc-framework-CoreMotion==10.3.1
pyobjc-framework-CoreServices==10.3.1
pyobjc-framework-CoreSpotlight==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-CoreWLAN==10.3.1
pyobjc-framework-CryptoTokenKit==10.3.1
pyobjc-framework-DataDetection==10.3.1
pyobjc-framework-DeviceCheck==10.3.1
pyobjc-framework-DictionaryServices==10.3.1
pyobjc-framework-DiscRecording==10.3.1
pyobjc-framework-DiscRecordingUI==10.3.1
pyobjc-framework-DiskArbitration==10.3.1
pyobjc-framework-DVDPlayback==10.3.1
pyobjc-framework-EventKit==10.3.1
pyobjc-framework-ExceptionHandling==10.3.1
pyobjc-framework-ExecutionPolicy==10.3.1
pyobjc-framework-ExtensionKit==10.3.1
pyobjc-framework-ExternalAccessory==10.3.1
pyobjc-framework-FileProvider==10.3.1
pyobjc-framework-FileProviderUI==10.3.1
pyobjc-framework-FinderSync==10.3.1
pyobjc-framework-FSEvents==10.3.1
pyobjc-framework-GameCenter==10.3.1
pyobjc-framework-GameController==10.3.1
pyobjc-framework-GameKit==10.3.1
pyobjc-framework-GameplayKit==10.3.1
pyobjc-framework-HealthKit==10.3.1
pyobjc-framework-ImageCaptureCore==10.3.1
pyobjc-framework-InputMethodKit==10.3.1
pyobjc-framework-InstallerPlugins==10.3.1
pyobjc-framework-InstantMessage==10.3.1
pyobjc-framework-Intents==10.3.1
pyobjc-framework-IntentsUI==10.3.1
pyobjc-framework-IOBluetooth==10.3.1
pyobjc-framework-IOBluetoothUI==10.3.1
pyobjc-framework-IOSurface==10.3.1
pyobjc-framework-iTunesLibrary==10.3.1
pyobjc-framework-KernelManagement==10.3.1
pyobjc-framework-LatentSemanticMapping==10.3.1
pyobjc-framework-LaunchServices==10.3.1
pyobjc-framework-libdispatch==10.3.1
pyobjc-framework-libxpc==10.3.1
pyobjc-framework-LinkPresentation==10.3.1
pyobjc-framework-LocalAuthentication==10.3.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.3.1
pyobjc-framework-MailKit==10.3.1
pyobjc-framework-MapKit==10.3.1
pyobjc-framework-MediaAccessibility==10.3.1
pyobjc-framework-MediaLibrary==10.3.1
pyobjc-framework-MediaPlayer==10.3.1
pyobjc-framework-MediaToolbox==10.3.1
pyobjc-framework-Metal==10.3.1
pyobjc-framework-MetalFX==10.3.1
pyobjc-framework-MetalKit==10.3.1
pyobjc-framework-MetalPerformanceShaders==10.3.1
pyobjc-framework-MetalPerformanceShadersGraph==10.3.1
pyobjc-framework-MetricKit==10.3.1
pyobjc-framework-MLCompute==10.3.1
pyobjc-framework-ModelIO==10.3.1
pyobjc-framework-MultipeerConnectivity==10.3.1
pyobjc-framework-NaturalLanguage==10.3.1
pyobjc-framework-NetFS==10.3.1
pyobjc-framework-Network==10.3.1
pyobjc-framework-NetworkExtension==10.3.1
pyobjc-framework-NotificationCenter==10.3.1
pyobjc-framework-OpenDirectory==10.3.1
pyobjc-framework-OSAKit==10.3.1
pyobjc-framework-OSLog==10.3.1
pyobjc-framework-PassKit==10.3.1
pyobjc-framework-PencilKit==10.3.1
pyobjc-framework-PHASE==10.3.1
pyobjc-framework-Photos==10.3.1
pyobjc-framework-PhotosUI==10.3.1
pyobjc-framework-PreferencePanes==10.3.1
pyobjc-framework-PushKit==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-QuickLookThumbnailing==10.3.1
pyobjc-framework-ReplayKit==10.3.1
pyobjc-framework-SafariServices==10.3.1
pyobjc-framework-SafetyKit==10.3.1
pyobjc-framework-SceneKit==10.3.1
pyobjc-framework-ScreenCaptureKit==10.3.1
pyobjc-framework-ScreenSaver==10.3.1
pyobjc-framework-ScreenTime==10.3.1
pyobjc-framework-ScriptingBridge==10.3.1
pyobjc-framework-SearchKit==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-SecurityFoundation==10.3.1
pyobjc-framework-SecurityInterface==10.3.1
pyobjc-framework-SensitiveContentAnalysis==10.3.1
pyobjc-framework-ServiceManagement==10.3.1
pyobjc-framework-SharedWithYou==10.3.1
pyobjc-framework-SharedWithYouCore==10.3.1
pyobjc-framework-ShazamKit==10.3.1
pyobjc-framework-Social==10.3.1
pyobjc-framework-SoundAnalysis==10.3.1
pyobjc-framework-Speech==10.3.1
pyobjc-framework-SpriteKit==10.3.1
pyobjc-framework-StoreKit==10.3.1
pyobjc-framework-Symbols==10.3.1
pyobjc-framework-SyncServices==10.3.1
pyobjc-framework-SystemConfiguration==10.3.1
pyobjc-framework-SystemExtensions==10.3.1
pyobjc-framework-ThreadNetwork==10.3.1
pyobjc-framework-UniformTypeIdentifiers==10.3.1
pyobjc-framework-UserNotifications==10.3.1
pyobjc-framework-UserNotificationsUI==10.3.1
pyobjc-framework-VideoSubscriberAccount==10.3.1
pyobjc-framework-VideoToolbox==10.3.1
pyobjc-framework-Virtualization==10.3.1
pyobjc-framework-Vision==10.3.1
pyobjc-framework-WebKit==10.3.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.6.3
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wavio==0.0.9
wcwidth==0.2.13
widgetsnbextension==4.0.11
yarl==1.9.4
